+++ mail_activity_team/README.rst
@@ -14,13 +14,13 @@ Mail Activity Team
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsocial-lightgray.png?logo=github
-    :target: https://github.com/OCA/social/tree/14.0/mail_activity_team
+    :target: https://github.com/OCA/social/tree/15.0/mail_activity_team
     :alt: OCA/social
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/social-14-0/social-14-0-mail_activity_team
+    :target: https://translation.odoo-community.org/projects/social-15-0/social-15-0-mail_activity_team
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
-    :target: https://runbot.odoo-community.org/runbot/205/14.0
+    :target: https://runbot.odoo-community.org/runbot/205/15.0
     :alt: Try me on Runbot
 
 |badge1| |badge2| |badge3| |badge4| |badge5| 
@@ -58,13 +58,18 @@ You can report on the activities assigned to a team going to
 *Dashboards / Activities*, and then filter by a specific team or group by
 teams.
 
+Known issues / Roadmap
+======================
+
+* In next version systray.esm.js file should be changed as added in comments.
+
 Bug Tracker
 ===========
 
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/social/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/social/issues/new?body=module:%20mail_activity_team%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/social/issues/new?body=module:%20mail_activity_team%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -74,17 +79,18 @@ Credits
 Authors
 ~~~~~~~
 
-* Eficent
+* ForgeFlow
 * Sodexis
 
 Contributors
 ~~~~~~~~~~~~
 
-* `Eficent <https://www.eficent.com>`_:
+* `ForgeFlow <https://www.forgeflow.com>`_:
 
   * Jordi Ballester Alomar (jordi.ballester@eficent.com)
   * Miquel Raïch (miquel.raich@eficent.com)
-  * Pedro Gonzalez (pedro.gonzalez@pesol.es)
+  * Bernat Puig Font (bernat.puig@forgeflow.com)
+* Pedro Gonzalez (pedro.gonzalez@pesol.es)
 * `Tecnativa <https://www.tecnativa.com>`_:
 
   * David Vidal
@@ -102,6 +108,6 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/social <https://github.com/OCA/social/tree/14.0/mail_activity_team>`_ project on GitHub.
+This module is part of the `OCA/social <https://github.com/OCA/social/tree/15.0/mail_activity_team>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
