+++ mail_activity_team/i18n/mail_activity_team.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
+"Project-Id-Version: Odoo Server 15.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -56,26 +56,17 @@ msgid "Created on"
 msgstr ""
 
 #. module: mail_activity_team
-#: model:ir.model.fields,field_description:mail_activity_team.field_mail_activity__display_name
-#: model:ir.model.fields,field_description:mail_activity_team.field_mail_activity_mixin__display_name
 #: model:ir.model.fields,field_description:mail_activity_team.field_mail_activity_team__display_name
-#: model:ir.model.fields,field_description:mail_activity_team.field_res_users__display_name
 msgid "Display Name"
 msgstr ""
 
 #. module: mail_activity_team
-#: model:ir.model.fields,field_description:mail_activity_team.field_mail_activity__id
-#: model:ir.model.fields,field_description:mail_activity_team.field_mail_activity_mixin__id
 #: model:ir.model.fields,field_description:mail_activity_team.field_mail_activity_team__id
-#: model:ir.model.fields,field_description:mail_activity_team.field_res_users__id
 msgid "ID"
 msgstr ""
 
 #. module: mail_activity_team
-#: model:ir.model.fields,field_description:mail_activity_team.field_mail_activity____last_update
-#: model:ir.model.fields,field_description:mail_activity_team.field_mail_activity_mixin____last_update
 #: model:ir.model.fields,field_description:mail_activity_team.field_mail_activity_team____last_update
-#: model:ir.model.fields,field_description:mail_activity_team.field_res_users____last_update
 msgid "Last Modified on"
 msgstr ""
 
@@ -152,7 +143,8 @@ msgstr ""
 #. module: mail_activity_team
 #: code:addons/mail_activity_team/models/mail_activity.py:0
 #, python-format
-msgid "The assigned user %s is not member of the team %s."
+msgid ""
+"The assigned user %(user_name)s is not member of the team %(team_name)s."
 msgstr ""
 
 #. module: mail_activity_team
@@ -166,16 +158,7 @@ msgid "Users"
 msgstr ""
 
 #. module: mail_activity_team
-#: model:ir.model.fields,field_description:mail_activity_team.field_account_bank_statement_line__activity_team_user_ids
-#: model:ir.model.fields,field_description:mail_activity_team.field_account_journal__activity_team_user_ids
-#: model:ir.model.fields,field_description:mail_activity_team.field_account_move__activity_team_user_ids
-#: model:ir.model.fields,field_description:mail_activity_team.field_account_payment__activity_team_user_ids
-#: model:ir.model.fields,field_description:mail_activity_team.field_event_event__activity_team_user_ids
-#: model:ir.model.fields,field_description:mail_activity_team.field_event_registration__activity_team_user_ids
 #: model:ir.model.fields,field_description:mail_activity_team.field_mail_activity_mixin__activity_team_user_ids
-#: model:ir.model.fields,field_description:mail_activity_team.field_mailing_mailing__activity_team_user_ids
-#: model:ir.model.fields,field_description:mail_activity_team.field_product_product__activity_team_user_ids
-#: model:ir.model.fields,field_description:mail_activity_team.field_product_template__activity_team_user_ids
 #: model:ir.model.fields,field_description:mail_activity_team.field_res_partner__activity_team_user_ids
 #: model:ir.model.fields,field_description:mail_activity_team.field_res_users__activity_team_user_ids
 msgid "test field"
