+++ mail_activity_team/models/mail_activity_team.py
@@ -1,4 +1,4 @@
-# Copyright 2018 Eficent Business and IT Consulting Services, S.L.
+# Copyright 2018-22 ForgeFlow Business and IT Consulting Services, S.L.
 # Copyright 2021 Sodexis
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 from odoo import api, fields, models
@@ -19,8 +19,8 @@ class MailActivityTeam(models.Model):
                 domain.append(("res_model_id", "in", team.res_model_ids.ids))
             team.count_missing_activities = activity_model.search(domain, count=True)
 
-    name = fields.Char(string="Name", required=True, translate=True)
-    active = fields.Boolean(string="Active", default=True)
+    name = fields.Char(required=True, translate=True)
+    active = fields.Boolean(default=True)
     res_model_ids = fields.Many2many(
         comodel_name="ir.model",
         string="Used models",
