+++ mail_activity_team/readme/CONTRIBUTORS.rst
@@ -1,8 +1,9 @@
-* `Eficent <https://www.eficent.com>`_:
+* `ForgeFlow <https://www.forgeflow.com>`_:
 
   * Jordi Ballester Alomar (jordi.ballester@eficent.com)
   * Miquel Raïch (miquel.raich@eficent.com)
-  * Pedro Gonzalez (pedro.gonzalez@pesol.es)
+  * Bernat Puig Font (bernat.puig@forgeflow.com)
+* Pedro Gonzalez (pedro.gonzalez@pesol.es)
 * `Tecnativa <https://www.tecnativa.com>`_:
 
   * David Vidal
